Tighten things up
